requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
attrs==25.4.0
audioread==3.1.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comtypes==1.4.15
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
edge-tts==7.2.7
einops==0.8.2
faiss-cpu==1.13.2
ffmpeg-python==0.2.0
filelock==3.20.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
future==1.0.0
gTTS==2.5.4
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.5.0
hyper-connections==0.4.9
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
local-attention==1.11.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
networkx==3.6.1
noisereduce==3.0.3
numba==0.63.1
numpy==2.3.5
ollama==0.6.1
omegaconf==2.3.0
packaging==26.0
pedalboard==0.9.22
pillow==12.1.0
platformdirs==4.5.1
pooch==1.9.0
praat-parselmouth==0.4.7
propcache==0.4.1
protobuf==7.34.0
PyAudio==0.2.14
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
pygame==2.6.1
Pygments==2.19.2
pyparsing==3.3.2
pypiwin32==223
python-dateutil==2.9.0.post0
pyttsx3==2.99
pywin32==311
pyworld==0.3.5
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
resampy==0.4.3
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.3
shellingham==1.5.4
six==1.17.0
sounddevice==0.5.5
soundfile==0.13.1
soxr==1.0.0
SpeechRecognition==3.14.5
stftpitchshift==2.0
sympy==1.14.0
tabulate==0.9.0
tensorboardX==2.6.4
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
torch-einops-utils==0.0.30
torchaudio==2.10.0
torchcrepe==0.0.24
torchfcpe==0.0.4
tqdm==4.67.3
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wget==3.2
yarl==1.22.0
